+++ grid2op/Action/_backendAction.py
@@ -240,6 +240,8 @@ class _BackendAction(GridObjects):
             self.shunt_p = ValueStore(self.n_shunt, dtype=dt_float)
             self.shunt_q = ValueStore(self.n_shunt, dtype=dt_float)
             self.shunt_bus = ValueStore(self.n_shunt, dtype=dt_int)
+            self.current_shunt_bus = ValueStore(self.n_shunt, dtype=dt_int)
+            self.current_shunt_bus.values[:] = 1
 
         self._status_or_before = np.ones(self.n_line, dtype=dt_int)
         self._status_ex_before = np.ones(self.n_line, dtype=dt_int)
@@ -268,6 +270,7 @@ class _BackendAction(GridObjects):
             res.shunt_p.copy(self.shunt_p)
             res.shunt_q.copy(self.shunt_q)
             res.shunt_bus.copy(self.shunt_bus)
+            res.current_shunt_bus.copy(self.current_shunt_bus)
 
         res._status_or_before[:] = self._status_or_before
         res._status_ex_before[:] = self._status_ex_before
@@ -308,6 +311,7 @@ class _BackendAction(GridObjects):
             self.shunt_p.reorder(no_shunt)
             self.shunt_q.reorder(no_shunt)
             self.shunt_bus.reorder(no_shunt)
+            self.current_shunt_bus.reorder(no_shunt)
 
     def reset(self):
         # last topo
@@ -331,6 +335,7 @@ class _BackendAction(GridObjects):
             self.shunt_p.reset()
             self.shunt_q.reset()
             self.shunt_bus.reset()
+            self.current_shunt_bus.reset()
 
     def all_changed(self):
         # last topo
@@ -404,7 +409,7 @@ class _BackendAction(GridObjects):
             self.storage_power.set_val(storage_power)
 
         # II shunts
-        if self.shunts_data_available:
+        if type(self).shunts_data_available:
             shunts = {}
             if other.shunts_data_available:
                 shunts["shunt_p"] = other.shunt_p
@@ -417,6 +422,7 @@ class _BackendAction(GridObjects):
             self.shunt_q.set_val(arr_)
             arr_ = shunts["shunt_bus"]
             self.shunt_bus.set_val(arr_)
+            self.current_shunt_bus.values[self.shunt_bus.changed] = self.shunt_bus.values[self.shunt_bus.changed]
 
         # III line status
         # this need to be done BEFORE the topology, as a connected powerline will be connected to their old bus.
@@ -575,9 +581,13 @@ class _BackendAction(GridObjects):
 
     def _get_active_bus(self):
         self.activated_bus[:, :] = False
-        tmp = self.current_topo.values - 1
+        tmp = self.current_topo.values - 1  # TODO global to local !
         is_el_conn = tmp >= 0
         self.activated_bus[self.big_topo_to_subid[is_el_conn], tmp[is_el_conn]] = True
+        if type(self).shunts_data_available:
+            is_el_conn = self.current_shunt_bus.values >= 0
+            tmp = self.current_shunt_bus.values - 1
+            self.activated_bus[type(self).shunt_to_subid[is_el_conn], tmp[is_el_conn]] = True
 
     def update_state(self, powerline_disconnected):
         """